1.1 edit

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -14,7 +14,7 @@
 
 unsigned long int _pow_recursion(int x, int y)
 {
-	unsigned int sum;
+	unsigned long int sum;
 
 	if (y == 0)
 		return (1);
